requirements.txt
openai
python-dotenv
streamlit
matplotlib
jupyterlab
langchain
Langchain-community
langchain-ollama
dashscope
chromadb